backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "groq>=0.31.0",
    "langchain-groq>=0.3.7",
    "portia-sdk-python[google,mistral]>=0.7.0",
    "supabase>=2.18.1",
]
